ai_service/requirements.txt
# requirements.txt
insightface
numpy
pillow
aio_pika  # สำหรับเชื่อม RabbitMQ
onnxruntime
opencv-python
aiohttp

